refactor: 삭제 버튼 모듈화
