+++ src/components/mydocs/lower/PreviewContent.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react'
 import styled from 'styled-components'
 import ViewDetailsButton from '../ViewDetailsButton'
-import { cardColorStore, isDeleteStore, previewOpenStore } from '../../../store/store'
-import Swal from 'sweetalert2'
+import { cardColorStore, previewOpenStore } from '../../../store/store'
 import ReactMarkdown from 'react-markdown'
 import PalleteButton from '../PalleteButton'
+import DeleteButton from '../DeleteButton'
 
 const Container = styled.div<{ previewOpen: boolean }>`
   position: absolute;
@@ -150,28 +150,6 @@ const CreatedAt = styled.p<{ color: string }>`
   margin: 0;
 `
 
-// 추가 버튼 (팔레트, 삭제)
-const OptionalButton = styled.button`
-  background: none;
-  border: none;
-  cursor: pointer;
-  font-size: 1.2rem;
-  padding: 0.5rem 0.7rem;
-  margin-left: 1rem;
-  outline: none;
-  transition: background-color 0.2s ease; // 배경색의 전환 시간 설정
-  &:hover {
-    background-color: rgba(0, 0, 0, 0.05); // 호버 시 배경색 조금 더 밝게
-  }
-  &:active {
-    outline: none; // 클릭 시 테두리가 나타나지 않도록 설정
-    background-color: rgba(0, 0, 0, 0.1); // 클릭 시 배경색 더 밝게
-  }
-  &:focus {
-    outline: none; // 포커스 시 테두리가 나타나지 않도록 설정
-  }
-`
-
 interface PreviewContentProps {
   color?: string
   title: string
@@ -189,8 +167,7 @@ const PreviewContent: React.FC<PreviewContentProps> = ({
   repo,
   tags,
 }) => {
-  const { previewOpen, setPreviewOpen } = previewOpenStore()
-  const { setIsDelete } = isDeleteStore()
+  const { previewOpen } = previewOpenStore()
   const { cardColor, setCardColor } = cardColorStore((state) => ({
     cardColor: state.cardColor,
     setCardColor: state.setCardColor,
@@ -201,30 +178,13 @@ const PreviewContent: React.FC<PreviewContentProps> = ({
     setCardColor(color || 'rgba(0, 0, 0, 1)')
   }, [color])
 
-  // 삭제 핸들링
-  const handleDelete = () => {
-    Swal.fire({
-      title: 'Are you sure?',
-      text: "You won't be able to revert this",
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonText: 'Delete it',
-      cancelButtonText: 'Cancel',
-    }).then((result) => {
-      if (result.isConfirmed) {
-        setIsDelete(true)
-        setPreviewOpen(false)
-      }
-    })
-  }
-
   return (
     <Container previewOpen={previewOpen} onClick={(e) => e.stopPropagation()}>
       <Wrapper>
         <ContentArea>
           <ButtonsContainer>
             <PalleteButton />
-            <OptionalButton onClick={handleDelete}>🗑️</OptionalButton>
+            <DeleteButton />
           </ButtonsContainer>
           <UpperWrapper>
             <Repo>{repo}</Repo>
